[chroma-nuget] --> integration (v0.1.0-alpha-002) OK

+++ DotnetLlamaSharp/Services/Embeddings/ChromaService.cs
@@ -1,5 +1,7 @@
 ﻿using AutoMapper;
 using DocumentFormat.OpenXml.Wordprocessing;
+using Dotnet.Chroma.Repositories.Models;
+using Dotnet.Chroma.Repositories.Models.Metadata;
 using DotnetLlamaSharp.Domain.Models.Entities.Chroma;
 using DotnetLlamaSharp.Domain.Models.Enums;
 using DotnetLlamaSharp.Domain.Models.Primitives.Chroma;
@@ -176,6 +178,6 @@ namespace DotnetLlamaSharp.Services.Embeddings
             => await _fileMgmtService.GetAllCollections();
 
         public async Task<List<ChromaChatsCollection>> GetAllChatCollections()
-            => await _chatsRepo.CollectionsOf(EChunkType.CHAT);
+            => await _chatsRepo.CollectionsOf((int)EChunkType.CHAT);
     }
 }
